may have fixed annoying make clean bugs
discovered that i was reinvoking the make for any cleaning process, just in order to set the CLEAN flag.  that's really terribly broken.  luckily found the eval command which i can run inside a target to change the value of an existing environment variable.  perfect.  now in the base clam rules and not getting weird problems from using OTHER_CLEANS