added a games category
info overload script got a new category for games and other fun stuff.  a lot of my notes are about game progress, install issues, problems running, quests to look up, and other silly stuff.  it needs its own category.  it's debatable whether it should count against the backlog, but i guess if it's in there, i supposedly intend to do it and thus it represents a time synch.