smoot-tastic with the .no-checkin
now can add a .no-checkin file at the top of a repository (svn, git, etc) and if that file is found, then the repository will not be checked in or have changes reported by our tools.  this is crucial for checking out others repos that we don't intend to modify.