renamings and add of singly linked list
fixed name of "list" class to be doubly linked list, since it is that.  now kind of grossed out by doubly linked list's implementation; suggest overhauling it so dl list IS a node in the dl list, rather than just a manager of node objects.  also noting it has a gross iterator implementation, but it appears to be critical to the design of the class, so try to leave it intact.
added a very basic singly linked list in order to have the new cycle detection algorithm.  added a test for singly linked lists.
added definition for eclipse on gnu unix of __UNIX__, which calmed a lot of complaints about HOOPLE_MAIN and other memset/strncpy/etc methods.