new alias netpush
got tired of typing the same old parameters to rsync, or sometimes still accidentally calling rcp instead, so now have an alias called 'netpush' that just does an 'rsync -avz' to move files or folders to remote (or local) targets.  remember to use the syntax for the destination with the hostname, colon, and path.  for example: netpush myfile surya:.  (copies myfile to the machine surya in the home directory of the user).