removed concept of LIGHTWEIGHT_INIT
this flag used to keep aliases and functions out of the startup, but this is problematic; most of the time we want our own macros and functions to be visible, and they don't really cost a lot anyhow.  so, dropping the flag entirely, which is nice; cleans up a few conditional areas.