tasty revision to load feisty for full use
instead of piecemeal loading of certain files, like functions.sh, i have swerved over into the oncoming lane where we now just pull in the feisty meow startup process.  shouldn't take any extra time if it was already loaded or assets like aliases survived into the running shell.  but the great thing is now that i can stop referring to scripts with paths if needed (when in bash), because the functions and aliases i expect to be able to use are now available.  dovetails with earlier removal of lightweight init feature.