simpler revision for terminal title
stopped trying to use a stack in an array, since that doesn't pass down to sub-shells.  not sure how i convinced myself that this approach ever worked; it was fully developed with tasty array code, but totally didn't solve the problem.
new approach simple uses a normal variable, which can be exported to sub-shells.  seems to work fairly well now.