fixes for unfortunate log change in screen parms
the maintainers of screen added a required parameter to the -L flag on
screen, which breaks every existing usage of the otherwise excellent
screen tool.  so, now we're forced to just add this useless parameter,
when my screenrc already takes care of setting a log file.  this change
is not even reflected on the gnu site for screen, so there's no
documentation for it at the official site.  shame on you, whoever just
stuffed this change into the screen tool.