fixing this up for site avenger
had been using a system level path for the site storage /srv/users/... but that's not how we roll with site avenger.  instead, the script now uses $HOME/apps as the storage folder.  also, the creation of a 'public' folder under the site is problematic, since that should be a link for site avenger sites, so the public dir creation has been removed.