attempting to fix naive system updater
the updater running from cron kept getting jammed up, with multiple possibly running at the same time.  this seems to be due to questions about using existing or new config files.  a fix suggested on internet says that dpkg needs to be told not to do that, and to just use the existing ones.