cleaning up revision control tools
tried to move to a consistent naming scheme.  getem and checkin are the "grab the code" and "check the code in" operations.  so rgetem and rcheckin are the recursive versions of those that act on a whole hierarchy.  other scripts have been renamed for easier understanding of what the heck they do.  some junk scripts have been flushed out.