even more careful push and pull
this may working now; we are comparing the branch to the parent, and were definitely seeing those as different at one point.  but now with a healthy development branch, the parent and branch are the same.  so the rule should be, hopefully, (1) when parent is different than self, you pull from your parent but push to yourself and (2) when parent and self are same, just push and pull.  if that's not the right grokking of it, then i will need to rethink some things.