changes to make careful update only for push
we don't generally want a regular update to get stuff from just everywhere; we want it to update if there's stuff on our specific remote origin.  so this only does the multi branch cleanup process when we're about to push changes, since that's when we'll get boned if the remotes aren't right.