attempt 3 and getting pretty frustrated
the exclude docs are not really helping, or the feature is just really
not working.  trying now to use just the most simplistic pattern
possible, which is only the dir name.  stars in the pattern just seem to
get completely ignored somehow, or slashes make it fail, or some kind of
insanity that seemingly didn't used to happen.