another attempt to clean up superfluous merges
we want to stop seeing pointless merge messages in the commit log.  this should take care of it, although there can be a problem if the code stashed before pulling is *way* out of synch with the code being pulled.  that problem would afflict the merge also though, and it is better to find out as soon as possible rather than letting the snarl grow larger.