better handling of dir tests
this adds handling of bad returns from the find_apps_folder and test_apps_folder methods.  previously we were relying on them to bail out with an exit, but cannot do that when using them within shell aliases in same shell.  we need them to just return their status, and now any place using them checks the status.  where appropriate this causes that outer shell to exit, but now the switchto macro will not bail out if the user chooses quit.