added dereferencing of links during snarf
con: this will not preserve directories exactly as shown in the original location.  pro: this will actually back up the data pointed at by those links, so that the intended purpose of backing up one's data is actually met.
=> this change still needs testing on system with letsencrypt installed, so we can prove that those files are backed up.

further, prior "sdimi" grunt was just trying to say that the changes in that revision were automatically caused by eclipse.