new scripts for acting on hierarchies
act_on_tree runs an action on every folder that is one level down from a directory, where the default is to use the current directory.  so like "act_on_tree git pull" would update all the git repositories under the current dir.  there is a flag for changing the default location to act on.
act_on_apps is a site-avenger specialized version of act_on_tree that runs the command on directories under ~/apps which also happen to be site avenger projects (i.e. they have an avenger5 subdirectory just under the project directory).