dropped terminal title oracle; things are good now
we stopped trying to scrape the current terminal title using various types of window magic.  the change from unity and having a WINDOWID defined really screwed us in that; we used to be able to get the exact right title based on our WINDOWID, but once that's missing, there's no reliable way to get the terminal title unless the terminal is in the foreground and focused.  heck with that.
so we're back to our approach of stacking up terminal titles.  this works now, since we have a better approach.  only a simple variable is used, not an array, so we can pass this to subshells reliably.  and the save / restore feature has been kept intact, even though save expects a parameter as the title now.  it just uses the existing item on the top of the stack if no new title is saved, thus preserving the semantics of saving the title.
so far things are looking nice.  it will be nice to decouple this feature of feisty meow from OS changes and screwiness.