fixed annoying bug for test section manager
where the test would create a dir /feisty_meow if run as root, for storing the ini file.
do need to reevaluate the all users dir option; it makes no sense on linux, although at least we drive it by a simple environment variable.  is there even a unix equivalent to that concept of storing for all users?  not really, except sort of /etc dir.