adding example apps, fixing powerup issues
the new example apps are in nascent state but provide two demos, one of the google maps using a zipcode database and the other implementing a shared calendar view in google.
the site avenger shared support and all scripts have been updated to allow an empty app_dirname, such as when you are already in a directory and want to just say powerup.  this works with existing avenger5 folders already, but still needs some help to be generalized.