adding path support to spacemall
this is a lot more obnoxious than expected.  it turns out spacem has a bug when run against a file in a directory, where it cannot do the renaming properly.
the current state is not right yet, but it's better than it was in terms of features.  but this needs to work properly for both spacem and spacemall