cleaning out references to avenger5 text
we already have a variable for this word, so it shouldn't appear in any file except the defaults file.
one allowed exception is a script that acts on all site avenger folders, since that script doesn't use the rest of the site avenger scripting machinery and it's an appropriate usage for this particular script (act on all apps).