really close for simple powerups
it looks like this is working now for a powerup done in a project directory with no parameters, even if that project directory is a way deep subdirectory of a git repository rather than a stand-alone repository.  had to stop before fully powered up due to blah blah but need to test out fully and then remove debug noises.
version control library had a feature removed which was not being used and seemed dangerous anyhow.