working vendor record plus new system time update
the vendor record got purged of its insane missing line problem by uploading it to google and then downloading it again.  freaking weird.
the update_system_time script is a simple wrapper for rdate, which has been throwing obnoxious "problem: successful" kinds of messages for a while.  don't know if the weird success message is accompanied by a result value of zero, as in success, or not, but we'll find out by what this script says.  now in use on serene for time updates.