new name for function
this is a local /z comparator, which abbreviates to localz somehow.  it will check the local machine's /z folder items against an exemplar machine on the network.  there is a built in default so far for that machine, but it also changes, which isn't so reasonable.  but an exemplar machine name can be passed on the command line, so we're not put together totally backwards like some IT groups.