improved f76 vendor record post bank merges
the fallout 76 banks all got tied to the same balance, so it no longer makes sense to track the locations of all the 7 vendors separately.  now it's all combined on one sheet, which should take a lot longer to fill out (which is a good thing to save paper and pen hand).