updates to get mac OS working again
note that building on recent versions of mac os requires the "brew" tools, which provide the gnu versions of a bunch of apps.  if there are complaints about command line flags, then it's probably because brew hasn't been used to replace the default mac tool.
here are some links that can help set this up:
https://brew.sh
https://gist.github.com/skyzyx/3438280b18e4f7c490db8a2a2ca0b9da
https://medium.com/@timmykko/using-openssl-library-with-macos-sierra-7807cfd47892