vital fixes to sudo function
not sure the initial thrust is done yet, of having root not get a pile of wrong feisty meow variables when i sudo to it with -E flag, but this does at least get the shell options to not still expand for the one case we need.  we will fix that to be more specific to enable mac to participate, since mac does actually need aliases expanded.
this also crucially fixes the IMPORTED_XAUTH variable which may never have been working right before?  saw a multi-line answer on my host, so taking just the first line of output.  not sure that's right yet, but we can't have multiple lines when we add the things in xauth the way we do it now.