simplified sudo function
tossed out a lot of my changes after discovering the -i option to sudo, which starts a legitimate fresh shell of the type i needed.  now, the only weirdity i have to deal with is that the PATH variable still seems to be passed in from sudo even though i've asked for that fresh shell.  setting it to null prior to doing the sudo fixes that.  works a bit better now.