new script for hiding output on success
this script will take a command line to execute and redirect the
commands stdout and stderr to two files.  if the command does not return
an error code, then the output files are just deleted.  if there is an
error reported, then the two files are sent back to stdout and stderr as
they would normally have been, and an additional error message is sent
to stderr.