tectonic tribulations edition: host preparation
needed for a bit; just automates a lot of separate install instructions.
planning to support windows with apt-cyg and mac with brew, but those steps are kinda hosed still.
is_feisty_up lets a caller know if the system is initialized yet or not, which was needed in prep_feisty_host.
functions have a couple changes, esp. a better whichable that now returns an error code instead of just not echoing text.