welding in a virtual unix root for windoze
since windows is too dumb to have a top-level root folder, we need to rely on the virtualization environment that is getting us the features we rely on, which usually come from unix (or linux).  the toolsets used in the past for this include MSYS, Cygwin, and possibly others.  this release is starting to support Cygwin and others will be added as time permits.