updates lurching towards functionality
first attempt at virtual unix root is not going to work out due to interdependencies introduced between configuration and filesystem or vice-versa.  i warned about them in a comment, and they unfortunately became true.  we need to pull the virtual unix root up to a new level, hoist it into a config file that we can generate at build time and not depend on a lot of file ops.