psfind working to get windows pid, but...
this winpid is not working in ps tools to lookup a process.
need to dig into this further, but hopefully the zap_fallout script will
start working again.
may have to add a flag to psfind that tells it to return either cygwin pids
for use in ps or winpids for use in taskkill, if the former is actually a
usable pid value in ps.