harmonizing shebang for scripts
this starts using the 'env' tool to locate bash for us, so we don't have
hard-coded paths.  the hard-coding of those always bugged me, and this
is a better solution, assuming that the environment has 'bash' resolve
to the right version of bash.  that is an externality that we cannot
enforce or guarantee, so it's appropriate to give up on trying to do so.