more changes to support a FEISTY_MEOW_PERSONAL_HOME variable
this acts in place of $HOME for any custom purposes that we refer
to $HOME for, when there's a choice in the matter.  we can't change
things about the system, like ssh looking for configs in $HOME/.ssh
and we don't want to.  this is just for assets related to or used by
the user that they want dealt with properly under their specific home directory.
if it's not already set, then we default to $HOME for the value.