release 2.140.188
jump in version; somehow the config version number was way out of sync with releases.
hopefully will stay in sync now.
lots of changes, mostly minor in this release.
main big thing was dropping some jars that were only present as a convenience for running
java apps, but which were ancient and had vulnerabilities.  trying to get back to a "no binaries"
standpoint in the git repository.