need better behavior in picking photo dirs
mushing this into better shape gradually.  switched f76 to linux from windows once saw that i could, so now the save dir is different.
could use new save dir as a standard!  it makes more sense to link directly to photos, than to have to both link to a folder and then find the photos.