fixed terminal titler
it was hosing up non-interactive bash sessions due to trying to echo out in command line friendly way.  bahhh, bad idea.
fix now does not interfere in non-interactive sessions by checking special bash magic variables.  we needed to add that,
even beyond all the original checks on terminal type and prompts, because those were all coming in indicating that it
was an interactive shell when it was not.  but the magic check for that 'i' being present in the "$-" variable (yes,
that's a dollar and a dash) did the trick.