revised to add --ssl flag
saw my backups starting to be empty, and found that this needs a --ssl
flag to be happy, except that before i made the change, the backups had
started working again.  not sure wtf, except maybe i installed a missing
bit last time i was fuggling around with this stuff?