traced an issue with mirror mode
we will only mirror, aka delete, when the hierarchy on the target is one that already exists on the source.
this is a consequence of the necessary scooping of the "*" under a hierarchy to be copied to the target,
rather than trying to copy at the same height for the directory trees (which will lead to a basement folder being
created under the remote target basement, i.e. /media/fred/archivedrive/basement/basement, rather than
copying direct contents.  so for now, we are limited by our use of rsync for this, unless there's a mode to
do a direct directory copy rather than its current behavior.