cleans directories with different hierarchies
this new tool compare_whacker is super handy.
i have a newer scheme for managing my digital pictures and an older scheme.
the two schemes are incompatible (old one used the camera name as the main index
and the new scheme just uses the data from the picture metadata).
the compare whacker was written so i could clean up a redundant copy of my pictures
that were stored in the old scheme.  i needed to find all the picture names that were
the same across the two.
NOTE: some pictures re-use the filenames on the pictures, and this creates havoc with
this script.  use more precise hierarchies if you have lots of identical file names!